Redesigning landing page to have a dynamic log in/sign up
